Issue #11736: Trial and error - Fix testCacheClearVisibility
